requirements.txt
# requirements.txt
mne>=1.7.1
pyedflib>=0.1.37
numpy>=1.26.4
scipy>=1.13.1 --only-binary=all
pywavelets==1.6.0
scikit-learn==1.5.1
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
umap-learn==0.5.6
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
onnx==1.16.1
onnxruntime==1.17.3
pytorch-lightning==2.3.0
tqdm==4.66.4
pyyaml==6.0.1
opencv-python==4.10.0.84